requirements.txt
blinker==1.4
brotlipy==0.7.0
certifi==2022.12.7
cmake==3.25.0
contourpy==1.1.1
cycler==0.11.0
filelock==3.9.0
fonttools==4.42.1
imagecodecs==2023.3.16
imageio==2.31.4
importlib-resources==6.0.1
Jinja2==3.1.2
kiwisolver==1.4.5
lazy_loader==0.3
lit==15.0.7
matplotlib==3.7.3
mkl-service==2.4.0
mpmath==1.2.1
networkx==3.0
numpy==1.24.1
opencv-python==4.8.0.76
opt-einsum==3.3.0
Pillow==9.3.0
protobuf==3.19.6
pyasn1-modules==0.2.8
pyparsing==3.1.1
PyWavelets==1.4.1
requests==2.28.1
requests-oauthlib==1.3.0
scikit-image==0.21.0
scipy==1.10.1
sympy==1.11.1
tensorboard-plugin-wit==1.8.1
tifffile==2023.7.10
tqdm==4.66.1
triton==2.0.0
typing_extensions==4.4.0
urllib3==1.26.13
